+++ netx/resolver/fake_test.go
@@ -4,8 +4,9 @@ import (
 	"context"
 	"io"
 	"net"
-	"testing"
 	"time"
+
+	"github.com/ooni/probe-engine/atomicx"
 )
 
 type FakeDialer struct {
@@ -100,25 +101,32 @@ func (fe FakeEncoder) Encode(domain string, qtype uint16, padding bool) ([]byte,
 	return fe.Data, fe.Err
 }
 
-func TestUnitFakeResolverThatFails(t *testing.T) {
-	client := NewFakeResolverThatFails()
-	addrs, err := client.LookupHost(context.Background(), "www.google.com")
-	if err == nil {
-		t.Fatal("expected an error here")
-	}
-	if addrs != nil {
-		t.Fatal("expected nil here")
-	}
+type FakeResolver struct {
+	NumFailures *atomicx.Int64
+	Err         error
+	Result      []string
 }
 
-func TestUnitFakeResolverWithResult(t *testing.T) {
-	orig := []string{"10.0.0.1"}
-	client := NewFakeResolverWithResult(orig)
-	addrs, err := client.LookupHost(context.Background(), "www.google.com")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(orig) != len(addrs) || orig[0] != addrs[0] {
-		t.Fatal("not the result we expected")
-	}
+func NewFakeResolverThatFails() FakeResolver {
+	return FakeResolver{NumFailures: atomicx.NewInt64(), Err: errNotFound}
 }
+
+func NewFakeResolverWithResult(r []string) FakeResolver {
+	return FakeResolver{NumFailures: atomicx.NewInt64(), Result: r}
+}
+
+var errNotFound = &net.DNSError{
+	Err: "no such host",
+}
+
+func (c FakeResolver) LookupHost(ctx context.Context, hostname string) ([]string, error) {
+	if c.Err != nil {
+		if c.NumFailures != nil {
+			c.NumFailures.Add(1)
+		}
+		return nil, c.Err
+	}
+	return c.Result, nil
+}
+
+var _ Resolver = FakeResolver{}
